api/requirements.txt
django==3.1.7
djangorestframework
psycopg2-binary
Pillow
six
django-rest-swagger